move helper code to custom hook
